pyproject.toml
[project]
name = "movie-ticketing-agent-adk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.18.0",
    "litellm>=1.79.3",
]
